// src/types/tution-fee.tsx
type TutionFee = {
  maThongTinHocPhi: number;
  soTienHocPhiTheoQuyDinh: number;
  soTienPhaiDong: number;
  soTienDaDong: number;
  soTienDu: number;
  tenNganHangThanhToanHocPhi: string;
  thoiDiemThanhToanHocPhi: string;
  ghiChuBoSung: string;
  maThongTinHocPhiHocKyTruoc: number | null;
  maHocKyNamHoc: number;
  maSinhVien: number;
  hocKyNamHoc: string | null;
  inverseThongTinHocPhiHocKyTruoc: any;
  sinhVien: any;
  thongTinHocKyNamHoc: any;
  thongTinHocPhiHocKyTruoc: any;
};